refactor: move discv4 server from mod to new file
